# v-hormiga-app-main/src/documentation/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MIKNE API",
    "description": "API for MIKNE application",
    "contact": {
      "name": "Support Team",
      "email": "lazcano@icloud.com"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginApiData"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginApiData"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginApiData"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginApiData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/loginpagos": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginApiData"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginApiData"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginApiData"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginApiData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginApiData": {
        "type": "object",
        "properties": {
          "companyDB": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}